Refactor of Codebase
Started (almost) from scratch, to avoid clutter. Will re-integrate with other branches in the future.
- All checks can now be found in the "Checks" folder.
- Checks can now be dragged and dropped into this folder, as long as they have a 'Check' class that inherits from "FormalCheckInterface".
- In each check, the analyse_cursor() function will run for each cursor.
- C++ source files have been moved to 'C++_Code'.
- The analyse_cursor() function returns a list of Alerts, a new class found in 'Core'. Alerts contain information about something that has gone wrong. They have their own 'display()' method, which is repurposed from print_error().
